Split GameObjectManager
